+++ src/__experimental__/components/radio-button/radio-button-group.style.js
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const RadioButtonGroupStyle = styled.div`
+  ${({
+    inline
+  }) => inline && 'display: flex;'}
+`;
+
+export default RadioButtonGroupStyle;