+++ src/components/drag-and-drop/with-drag/__spec__.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { mount } from 'enzyme';
-import { DragDropContext } from 'react-dnd';
+import { DragDropContext } from 'react-dnd-legacy';
 import TestBackend from 'react-dnd-test-backend';
 import WithDrag from './with-drag';
 import BrowserHelper from './../../../utils/helpers/browser';